Refactor app.py to use blueprnts
There is now a blueprint for the home views (index, instructions,
enter code) and another for level vies (just the test level and
its refresh for now. This will make adding the functions for
levels much cleaner by doing it as a service that is imported
by the level views.