Thor fleshed out and large keys added
We now have Thor and his games pretty fleshed out. We
have a floor with two of them. The second gives a golden key.

The gold key can be used on a golden door!

The first floor is now recreated with silver keys on the floor
replacing Audumbla. That damn cow is next...