Made thor/cow games each one instance
Now you pass in the text and reward as arguments.
This guarantees that when you get to a certain tile
that creature is the same every time.
It also moves text to dungeon creation file which
makes so much more sense and I wanted this set up before
doing spooktober.